+++ PolarisEngine/src/Core/TimeStep.cpp
@@ -2,6 +2,8 @@
 
 #include <GLFW/glfw3.h>
 
+Time* Time::instance = nullptr;
+
 Time::Time() : deltaTime(0), lastFrameTime(0)
 {
 
@@ -13,12 +15,24 @@ Time::~Time()
 
 void Time::Init()
 {
+	instance = new Time();
+
+	instance->InitImplementation();
+}
+
+void Time::InitImplementation()
+{
 	glfwSetTime(0);
 
 	lastFrameTime = glfwGetTime();
 }
 
 void Time::Update()
+{
+	instance->UpdateImplementation();
+}
+
+void Time::UpdateImplementation()
 {
 	GLfloat currentTime = glfwGetTime();
 	deltaTime = currentTime - lastFrameTime;
@@ -27,10 +41,20 @@ void Time::Update()
 
 GLfloat Time::GetDeltaTime()
 {
+	return instance->GetDeltaTimeImplementation();
+}
+
+GLfloat Time::GetDeltaTimeImplementation()
+{
 	return deltaTime;
 }
 
 GLfloat Time::GetTimeSinceStart()
+{
+	return instance->GetDeltaTimeImplementation();
+}
+
+GLfloat Time::GetTimeSinceStartImplementation()
 {
 	return glfwGetTime();
 }